+++ CHANGELOG.md
@@ -2,15 +2,35 @@
 
 ## 1.0.0
 
-- [BREAKING] IEntity is now typed differently - it requires `typeof` modifier for components. Example:
+BREAKING CHANGES:
+
+- refactor: IEntity is now typed differently - it requires `typeof` modifier for components.
+- feat: All argument types are now correctly inferred for Components which use constructor args.
+- feat: `addComponent` now accepts arguments positionally rather than packed in an Array.
+
+Example for above changes:
 
 ```tsx
+  class FakeComponentWithArgs {
+    private _arg1: number;
+    private _arg2: string;
+
+    constructor(arg1: number, arg2: string) {
+      this._arg1 = arg1;
+      this._arg2 = arg2;
+    }
+  }
+
   let entity: IEntity<
     [
       typeof FakeComponentWithNoArgs,
       typeof FakeComponentWithArgs
     ]
   >;
+
+  (...)
+
+  entity.addComponent(FakeComponentWithArgs, 1, "test"); // this now correctly infers types.
 ```
 
 ## 0.3.0